fix: Fix CHAINID opcode chain ID source and type consistency
Changes:
- Changed EVM.get_chain_id() to return block_info.chain_id (u64) instead of context.chain_id (u16)
- Updated all mock implementations to return u64 consistently
- Fixed CALLCODE handler to use getEvm().inner_call instead of self.host.inner_call
- Fixed unused variable warning in frame.zig

The CHAINID opcode should return the EIP-155 chain ID from block info, which supports
larger chain IDs (like Polygon 137, BSC 56) that don't fit in u16.

+++ src/evm/frame_bench.zig
@@ -227,7 +227,7 @@ const TestHost = struct {
         return &[_]u8{};
     }
 
-    pub fn get_chain_id(self: *Self) u16 {
+    pub fn get_chain_id(self: *Self) u64 {
         _ = self;
         return 1;
     }
