+++ src/evm/frame_tests.zig
@@ -234,7 +234,7 @@ const MockEvm = struct {
         return &[_]u8{};
     }
 
-    pub fn get_chain_id(self: *Self) u16 {
+    pub fn get_chain_id(self: *Self) u64 {
         _ = self;
         return 1;
     }
